// src/lib/sdk/models/RegisterDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RegisterDto = {
    email: string;
    password: string;
    displayName?: string;
    /**
     * URL to avatar image
     */
    avatar?: string;
};